update auth structure

+++ twitter_clone-server/controller/tweet.js
@@ -38,6 +38,13 @@ export async function updateTweet (req, res) {
     try {
         const tweetId = req.params.id
         const text = req.body.text
+        const tweet = await tweetRepo.getById(tweetId)
+        if (!tweet) {
+            return res.sendStatus(404)
+        }
+        if (tweet.userId !== req.userId) {
+            return res.sendStatus(403)
+        }
         const updatedTweet = await tweetRepo.update(tweetId, text)
         res.status(200).json(updatedTweet)
     } catch (error) {
@@ -48,6 +55,13 @@ export async function updateTweet (req, res) {
 
 export async function deleteTweet (req, res) {
     try {
+        const tweet = await tweetRepo.getById(tweetId)
+        if (!tweet) {
+            return res.sendStatus(404)
+        }
+        if (tweet.userId !== req.userId) {
+            return res.sendStatus(403)
+        }
         await tweetRepo.remove(req.params.id)
         res.sendStatus(204)
     } catch (error) {
